Remove Deno.test from source file
