+++ src/func.ts
@@ -1,4 +1,3 @@
-import { assertEquals, assertThrows } from "../deps.ts";
 import type { Apply2Only, Hkt2 } from "./hkt.ts";
 import type { MonadReader } from "./reader/monad.ts";
 import { type AbelianGroup, abelSymbol } from "./type-class/abelian-group.ts";
@@ -8,7 +7,6 @@ import type { Functor } from "./type-class/functor.ts";
 import { type Group } from "./type-class/group.ts";
 import type { Monad } from "./type-class/monad.ts";
 import { semiGroupSymbol } from "./type-class/semi-group.ts";
-import { addAbelianGroup } from "./number.ts";
 
 /**
  * The type of function from `A` to `B`.
@@ -19,38 +17,53 @@ export interface Fn<A, B> {
 
 /**
  * The identity function which returns the passed value as is.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { assertEquals } from "../deps.ts";
+ * import { id } from "./func.ts";
+ *
+ * assertEquals(id(2), 2);
+ * assertEquals(id("foo"), "foo");
+ * ```
  */
 export const id = <T>(x: T) => x;
 
-Deno.test("id", () => {
-    assertEquals(id(2), 2);
-    assertEquals(id("foo"), "foo");
-});
-
 /**
  * The constant function which returns the first passed value.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { assertEquals } from "../deps.ts";
+ * import { constant } from "./func.ts";
+ *
+ * const fn = constant(4);
+ * assertEquals(fn(3), 4);
+ * assertEquals(fn("foo"), 4);
+ * ```
  */
 export const constant = <T>(x: T) => <U>(_u: U) => x;
 
-Deno.test("constant", () => {
-    const fn = constant(4);
-    assertEquals(fn(3), 4);
-    assertEquals(fn("foo"), 4);
-});
-
 /**
  * Indicates the unreachable code path. Calling this throws an error immediately, so this function should be called only if your application data is not consistent.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { absurd } from "./func.ts";
+ * import { assertThrows } from "../deps.ts";
+ *
+ * assertThrows(() => {
+ *     absurd<number>();
+ *     throw new Error("this line must not be run");
+ * }, "PANIC: absurd must not be called");
+ * ```
  */
 export const absurd = <T>(): T => {
     throw new Error("PANIC: absurd must not be called");
 };
-
-Deno.test("absurd", () => {
-    assertThrows(() => {
-        absurd<number>();
-        throw new Error("this line must not be run");
-    }, "PANIC: absurd must not be called");
-});
 
 /**
  * Composes two functions sequentially. `z = pipe(a)(b)(x)` means `y = a(x), z = b(y)`.
@@ -58,14 +71,19 @@ Deno.test("absurd", () => {
  * @param firstDo - The function to do at first.
  * @param secondDo - The function to do at last.
  * @returns The composed function.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { assertEquals } from "../deps.ts";
+ * import { pipe } from "./func.ts";
+ *
+ * assertEquals(pipe((x: number) => x + 1)((x) => x * 2)(3), 8);
+ * ```
  */
 export const pipe =
     <T, U>(firstDo: Fn<T, U>) => <V>(secondDo: Fn<U, V>) => (t: T) =>
         secondDo(firstDo(t));
-
-Deno.test("pipe", () => {
-    assertEquals(pipe((x: number) => x + 1)((x) => x * 2)(3), 8);
-});
 
 /**
  * Composes two functions mathematically. `compose(f)(g)(x)` means `f(g(x))`.
@@ -73,28 +91,38 @@ Deno.test("pipe", () => {
  * @param f - The function to do at last.
  * @param g - The function to do at first.
  * @returns The composed function.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { assertEquals } from "../deps.ts";
+ * import { compose } from "./func.ts";
+ *
+ * assertEquals(compose((x: number) => x + 1)((x: number) => x * 2)(3), 7);
+ * ```
  */
 export const compose = <U, V>(f: Fn<U, V>) => <T>(g: Fn<T, U>) => (t: T) =>
     f(g(t));
-
-Deno.test("compose", () => {
-    assertEquals(compose((x: number) => x + 1)((x: number) => x * 2)(3), 7);
-});
 
 /**
  * Flips two arguments of the function.
  *
  * @param f - The function with two arguments.
  * @returns The function flipped the arguments.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { assertEquals } from "../deps.ts";
+ * import { flip } from "./func.ts";
+ *
+ * const fn = flip((a: string) => (b: string) => a + b);
+ * assertEquals(fn("a")("b"), "ba");
+ * assertEquals(fn("asd")("btg"), "btgasd");
+ * ```
  */
 export const flip =
     <T, U, V>(f: Fn<T, Fn<U, V>>): Fn<U, Fn<T, V>> => (u) => (t) => f(t)(u);
-
-Deno.test("flip", () => {
-    const fn = flip((a: string) => (b: string) => a + b);
-    assertEquals(fn("a")("b"), "ba");
-    assertEquals(fn("asd")("btg"), "btgasd");
-});
 
 /**
  * Repeats the `succ` operation until `pred` returns `true`.
@@ -103,6 +131,21 @@ Deno.test("flip", () => {
  * @param succ - The recursion successor of the operation.
  * @param init - The initial term of the operation.
  * @returns The found value while operations.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { assertEquals } from "../deps.ts";
+ * import { until } from "./func.ts";
+ *
+ * const padLeft = until((x: string) => 4 <= x.length)((x) => "0" + x);
+ * assertEquals(padLeft(""), "0000");
+ * assertEquals(padLeft("1"), "0001");
+ * assertEquals(padLeft("13"), "0013");
+ * assertEquals(padLeft("131"), "0131");
+ * assertEquals(padLeft("1316"), "1316");
+ * assertEquals(padLeft("1316534"), "1316534");
+ * ```
  */
 export const until =
     <T>(pred: (t: T) => boolean) => (succ: (t: T) => T): (init: T) => T => {
@@ -115,30 +158,25 @@ export const until =
         return go;
     };
 
-Deno.test("until", () => {
-    const padLeft = until((x: string) => 4 <= x.length)((x) => "0" + x);
-    assertEquals(padLeft(""), "0000");
-    assertEquals(padLeft("1"), "0001");
-    assertEquals(padLeft("13"), "0013");
-    assertEquals(padLeft("131"), "0131");
-    assertEquals(padLeft("1316"), "1316");
-    assertEquals(padLeft("1316534"), "1316534");
-});
-
 /**
  * Maps the hom `X => A` with `f`.
  *
  * @param f - The function to map from `A`.
  * @param a - The hom to be mapped.
  * @returns The mapped hom.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { assertEquals } from "../deps.ts";
+ * import { map } from "./func.ts";
+ *
+ * const mapper = map<string>()((x: number) => x * 2);
+ * assertEquals(mapper(parseInt)("20"), 40);
+ * ```
  */
 export const map = <X>() => <A, B>(f: (a: A) => B) => (a: Fn<X, A>): Fn<X, B> =>
     pipe(a)(f);
-
-Deno.test("map", () => {
-    const mapper = map<string>()((x: number) => x * 2);
-    assertEquals(mapper(parseInt)("20"), 40);
-});
 
 /**
  * Applies the hom `X => A => B` to another hom `X => A`.
@@ -146,17 +184,22 @@ Deno.test("map", () => {
  * @param f - The hom returns the function to apply.
  * @param g - The hom to be applied.
  * @returns The applied hom.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { apply } from "./func.ts";
+ * import { assertEquals } from "../deps.ts";
+ *
+ * const applier = apply<string>()((str) => (radix: number) =>
+ *     parseInt(str, radix)
+ * );
+ * assertEquals(applier(parseInt)("11"), 12);
+ * ```
  */
 export const apply =
     <X>() => <A, B>(f: Fn<X, (a: A) => B>) => (g: Fn<X, A>): Fn<X, B> => (x) =>
         f(x)(g(x));
-
-Deno.test("apply", () => {
-    const applier = apply<string>()((str) => (radix: number) =>
-        parseInt(str, radix)
-    );
-    assertEquals(applier(parseInt)("11"), 12);
-});
 
 /**
  * Lifts the binary operation `q` over the hom `X => _`.
@@ -165,6 +208,16 @@ Deno.test("apply", () => {
  * @param f - The hom `X => A`.
  * @param g - The hom `X => B`.
  * @returns The lifted hom `X => C`.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { assertEquals } from "../deps.ts";
+ * import { liftBinary } from "./func.ts";
+ *
+ * const lifter = liftBinary<void>()((a: number) => (b: number) => a + b);
+ * assertEquals(lifter(() => 1)(() => 2)(), 3);
+ * ```
  */
 export const liftBinary =
     <X>() =>
@@ -173,26 +226,26 @@ export const liftBinary =
     (g: Fn<X, B>): Fn<X, C> =>
     (x) => q(f(x))(g(x));
 
-Deno.test("liftBinary", () => {
-    const lifter = liftBinary<void>()((a: number) => (b: number) => a + b);
-    assertEquals(lifter(() => 1)(() => 2)(), 3);
-});
-
 /**
  * Maps and flattens the hom with `fn`.
  *
  * @param fn - The function which maps from `A` to the hom `X => B`.
  * @param a - The hom `X => A`.
  * @returns The mapped hom `X => B`.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { assertEquals } from "../deps.ts";
+ * import { flatMap } from "./func.ts";
+ *
+ * const mapper = flatMap<number>()((x: number) => (y: number) => x * y);
+ * assertEquals(mapper((x) => x + 1)(3), 12);
+ * ```
  */
 export const flatMap =
     <X>() => <A, B>(fn: (a: A) => Fn<X, B>) => (a: Fn<X, A>): Fn<X, B> => (x) =>
         fn(a(x))(x);
-
-Deno.test("flatMap", () => {
-    const mapper = flatMap<number>()((x: number) => (y: number) => x * y);
-    assertEquals(mapper((x) => x + 1)(3), 12);
-});
 
 export interface FnHkt extends Hkt2 {
     readonly type: Fn<this["arg2"], this["arg1"]>;
@@ -264,19 +317,4 @@ export const abelianGroup = <A, B>(
     invert: (g) => (a) => gr.invert(g(a)),
     [semiGroupSymbol]: true,
     [abelSymbol]: true,
-});
-
-Deno.test("group", () => {
-    for (const make of [group, abelianGroup]) {
-        const groupForFn = make<void, number>(addAbelianGroup);
-        assertEquals(groupForFn.identity(), 0);
-        assertEquals(
-            groupForFn.combine(
-                () => 1,
-                () => 2,
-            )(),
-            3,
-        );
-        assertEquals(groupForFn.invert(() => 1)(), -1);
-    }
 });
