+++ src/number.ts
@@ -1,4 +1,3 @@
-import { assertEquals } from "../deps.ts";
 import { none, type Option, some } from "./option.ts";
 import { equal, greater, less, type Ordering } from "./ordering.ts";
 import {
@@ -24,16 +23,6 @@ export const partialCmp = (lhs: number, rhs: number): Option<Ordering> => {
     return some(greater);
 };
 export const partialOrd = fromPartialCmp(() => partialCmp)();
-
-Deno.test("partialCmp", () => {
-    assertEquals(partialCmp(1, NaN), none());
-    assertEquals(partialCmp(NaN, 2), none());
-    assertEquals(partialCmp(NaN, NaN), none());
-    assertEquals(partialCmp(1, 1), some(equal as Ordering));
-    assertEquals(partialCmp(2, 2), some(equal as Ordering));
-    assertEquals(partialCmp(1, 2), some(less as Ordering));
-    assertEquals(partialCmp(2, 1), some(greater as Ordering));
-});
 
 export const addAbelianGroup: AbelianGroup<number> = {
     combine: (l, r) => l + r,
