+++ src/result.ts
@@ -1,4 +1,3 @@
-import { assertEquals, assertThrows } from "../deps.ts";
 import type { Get1, Hkt2 } from "./hkt.ts";
 import type { Optic } from "./optical.ts";
 import { newPrism } from "./optical/prism.ts";
@@ -63,14 +62,19 @@ export const err = <E>(e: E): Err<E> => [errSymbol, e];
  *
  * @param res - The result to be checked.
  * @returns Whether the result is an `Ok`.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { err, isOk, ok } from "./result.ts";
+ * import { assertEquals } from "../deps.ts";
+ *
+ * assertEquals(isOk(ok(-3)), true);
+ * assertEquals(isOk(err("Some error message")), false);
+ * ```
  */
 export const isOk = <E, T>(res: Result<E, T>): res is Ok<T> =>
     res[0] === okSymbol;
-
-Deno.test("isOk", () => {
-    assertEquals(isOk(ok(-3)), true);
-    assertEquals(isOk(err("Some error message")), false);
-});
 
 /**
  * Checks whether the result is an `Err`.
@@ -80,11 +84,6 @@ Deno.test("isOk", () => {
  */
 export const isErr = <E, T>(res: Result<E, T>): res is Err<E> =>
     res[0] === errSymbol;
-
-Deno.test("isErr", () => {
-    assertEquals(isErr(ok(-3)), false);
-    assertEquals(isErr(err("Some error message")), true);
-});
 
 export const partialEquality = <E, T>(
     { equalityE, equalityT }: {
@@ -159,35 +158,55 @@ export const either =
  *
  * @param resRes - The nested result.
  * @returns The unwrapped nest.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { err, flatten, ok } from "./result.ts";
+ * import { assertEquals } from "../deps.ts";
+ *
+ * assertEquals(flatten(ok(ok("hello"))), ok("hello"));
+ * assertEquals(flatten(err(ok("hello"))), err(ok("hello")));
+ * assertEquals(flatten(ok(err(6))), err(6));
+ * assertEquals(flatten(err(err(6))), err(err(6)));
+ * ```
  */
 export const flatten = <E, T>(resRes: Result<E, Result<E, T>>): Result<E, T> =>
     isOk(resRes) ? resRes[1] : err(resRes[1]);
-
-Deno.test("flatten", () => {
-    assertEquals(flatten(ok(ok("hello"))), ok("hello"));
-    assertEquals(flatten(err(ok("hello"))), err(ok("hello")));
-    assertEquals(flatten(ok(err(6))), err(6));
-    assertEquals(flatten(err(err(6))), err(err(6)));
-});
 
 /**
  * Unwraps the value for the result whose type parameters of are same.
  *
  * @param res - The source result.
  * @returns The contained value.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { err, mergeOkErr, ok } from "./result.ts";
+ * import { assertEquals } from "../deps.ts";
+ *
+ * assertEquals(mergeOkErr(ok(3)), 3);
+ * assertEquals(mergeOkErr(err(4)), 4);
+ * ```
  */
 export const mergeOkErr = <T>(res: Result<T, T>) => res[1];
-
-Deno.test("mergeOkErr", () => {
-    assertEquals(mergeOkErr(ok(3)), 3);
-    assertEquals(mergeOkErr(err(4)), 4);
-});
 
 /**
  * Unwraps the `Ok` value from a `Result`, or throws an error.
  *
  * @param res - The value which should be an `Ok`.
  * @returns The unwrapped item.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { err, ok, unwrap } from "./result.ts";
+ * import { assertEquals, assertThrows } from "../deps.ts";
+ *
+ * assertEquals(unwrap(ok(3)), 3);
+ * assertThrows(() => unwrap(err(4)), "unwrapped Err");
+ * ```
  */
 export const unwrap = <E, T>(res: Result<E, T>): T => {
     if (isErr(res)) {
@@ -196,16 +215,21 @@ export const unwrap = <E, T>(res: Result<E, T>): T => {
     return res[1];
 };
 
-Deno.test("unwrap", () => {
-    assertEquals(unwrap(ok(3)), 3);
-    assertThrows(() => unwrap(err(4)), "unwrapped Err");
-});
-
 /**
  * Unwraps the `Err` value from a `Result`, or throws an error.
  *
  * @param res - The value which should be an `Err`.
  * @returns The unwrapped item.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { err, ok, unwrapErr } from "./result.ts";
+ * import { assertEquals, assertThrows } from "../deps.ts";
+ *
+ * assertThrows(() => unwrapErr(ok(3)), "unwrapped Ok");
+ * assertEquals(unwrapErr(err(4)), 4);
+ * ```
  */
 export const unwrapErr = <E, T>(res: Result<E, T>): E => {
     if (isOk(res)) {
@@ -214,33 +238,33 @@ export const unwrapErr = <E, T>(res: Result<E, T>): E => {
     return res[1];
 };
 
-Deno.test("unwrapErr", () => {
-    assertThrows(() => unwrapErr(ok(3)), "unwrapped Ok");
-    assertEquals(unwrapErr(err(4)), 4);
-});
-
 /**
  * Returns `resB` if `resA` is an `Ok`, otherwise returns the error `resA`. The order of arguments is reversed because of that it is useful for partial applying.
  *
  * @param resB - The second result.
  * @param resA - The first result.
  * @returns `resB` if `resA` is a `Ok`.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { and, err, ok } from "./result.ts";
+ * import { assertEquals } from "../deps.ts";
+ *
+ * const success = ok<number>(2);
+ * const failure = err("not a 2");
+ * const lateError = err("late error");
+ * const earlyError = err("early error");
+ * const anotherSuccess = ok("different result");
+
+ * assertEquals(and(lateError)(success), lateError);
+ * assertEquals(and<number, string>(success)(earlyError), earlyError);
+ * assertEquals(and(lateError)(failure), failure);
+ * assertEquals(and(anotherSuccess)(success), anotherSuccess);
+ * ```
  */
 export const and = <U, E>(resB: Result<E, U>) => <T>(resA: Result<E, T>) =>
     isOk(resA) ? resB : resA;
-
-Deno.test("and", () => {
-    const success = ok<number>(2);
-    const failure = err("not a 2");
-    const lateError = err("late error");
-    const earlyError = err("early error");
-    const anotherSuccess = ok("different result");
-
-    assertEquals(and(lateError)(success), lateError);
-    assertEquals(and<number, string>(success)(earlyError), earlyError);
-    assertEquals(and(lateError)(failure), failure);
-    assertEquals(and(anotherSuccess)(success), anotherSuccess);
-});
 
 /**
  * Returns `fn()` if `resA` is an `Ok`, otherwise returns the error `resA`. This is an implementation of `FlatMap`. The order of arguments is reversed because of that it is useful for partial applying.
@@ -248,23 +272,28 @@ Deno.test("and", () => {
  * @param fn - The function provides a second result.
  * @param resA - The first result.
  * @returns `fn()` if `resA` is an `Ok`.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { andThen, err, ok, Result } from "./result.ts";
+ * import { assertEquals } from "../deps.ts";
+ *
+ * const sqrtThenToString = andThen(
+ *     (num: number): Result<string, string> =>
+ *         num < 0
+ *             ? err("num must not be negative")
+ *             : ok(Math.sqrt(num).toString()),
+ * );
+
+ * assertEquals(sqrtThenToString(ok(4)), ok("2"));
+ * assertEquals(sqrtThenToString(ok(-1)), err("num must not be negative"));
+ * assertEquals(sqrtThenToString(err("not a number")), err("not a number"));
+ * ```
  */
 export const andThen =
     <T, U, E>(fn: (t: T) => Result<E, U>) => (resA: Result<E, T>) =>
         isOk(resA) ? fn(resA[1]) : resA;
-
-Deno.test("andThen", () => {
-    const sqrtThenToString = andThen(
-        (num: number): Result<string, string> =>
-            num < 0
-                ? err("num must not be negative")
-                : ok(Math.sqrt(num).toString()),
-    );
-
-    assertEquals(sqrtThenToString(ok(4)), ok("2"));
-    assertEquals(sqrtThenToString(ok(-1)), err("num must not be negative"));
-    assertEquals(sqrtThenToString(err("not a number")), err("not a number"));
-});
 
 /**
  * Returns `resB` if `resA` is an `Err`, otherwise returns the success `resA`. The order of arguments is reversed because of that it is useful for partial applying.
@@ -272,22 +301,27 @@ Deno.test("andThen", () => {
  * @param resB - The second result.
  * @param resA - The first result.
  * @returns `resA` or `resB`.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { err, ok, or } from "./result.ts";
+ * import { assertEquals } from "../deps.ts";
+ *
+ * const success = ok<number>(2);
+ * const failure = err<string>("not a 2");
+ * const lateError = err<string>("late error");
+ * const earlyError = err<string>("early error");
+ * const anotherSuccess = ok<number>(100);
+
+ * assertEquals(or<string, number>(lateError)(success), success);
+ * assertEquals(or<string, number>(success)(earlyError), success);
+ * assertEquals(or(lateError)(failure), lateError);
+ * assertEquals(or(anotherSuccess)(success), success);
+ * ```
  */
 export const or = <E, T>(resB: Result<E, T>) => (resA: Result<E, T>) =>
     isErr(resA) ? resB : resA;
-
-Deno.test("or", () => {
-    const success = ok<number>(2);
-    const failure = err<string>("not a 2");
-    const lateError = err<string>("late error");
-    const earlyError = err<string>("early error");
-    const anotherSuccess = ok<number>(100);
-
-    assertEquals(or<string, number>(lateError)(success), success);
-    assertEquals(or<string, number>(success)(earlyError), success);
-    assertEquals(or(lateError)(failure), lateError);
-    assertEquals(or(anotherSuccess)(success), success);
-});
 
 /**
  * Returns `fn()` if `resA` is an `Err`, otherwise returns the success `resA`. The order of arguments is reversed because of that it is useful for partial applying.
@@ -295,77 +329,107 @@ Deno.test("or", () => {
  * @param fn - The second result.
  * @param resA - The first result.
  * @returns `resA` or `fn()`.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { err, ok, orElse } from "./result.ts";
+ * import { assertEquals } from "../deps.ts";
+ *
+ * const sq = orElse((x: number) => ok<number>(x * x));
+ * const residual = orElse((x: number) => err<number>(x));
+
+ * assertEquals(sq(sq(ok(2))), ok(2));
+ * assertEquals(sq(residual(ok(2))), ok(2));
+ * assertEquals(residual(sq(err(3))), ok(9));
+ * assertEquals(residual(residual(err(3))), err(3));
+ * ```
  */
 export const orElse =
     <E, T, F>(fn: (error: E) => Result<F, T>) => (resA: Result<E, T>) =>
         isErr(resA) ? fn(resA[1]) : resA;
-
-Deno.test("orElse", () => {
-    const sq = orElse((x: number) => ok<number>(x * x));
-    const residual = orElse((x: number) => err<number>(x));
-
-    assertEquals(sq(sq(ok(2))), ok(2));
-    assertEquals(sq(residual(ok(2))), ok(2));
-    assertEquals(residual(sq(err(3))), ok(9));
-    assertEquals(residual(residual(err(3))), err(3));
-});
 
 /**
  * Takes the success value as an optional if the result is an `Ok`, otherwise returns `None`.
  *
  * @param res - The source result.
  * @returns The success value if exists.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { err, ok, optionOk, orElse } from "./result.ts";
+ * import { some, none } from "./option.ts";
+ * import { assertEquals } from "../deps.ts";
+ *
+ * const sq = orElse((x: number) => ok<number>(x * x));
+ * const residual = orElse((x: number) => err<number>(x));
+ *
+ * assertEquals(optionOk(ok(2)), some(2));
+ * assertEquals(optionOk(err("nothing left")), none());
+ * ```
  */
 export const optionOk = <E, T>(
     res: Result<E, T>,
 ): Option<T> => (isOk(res) ? some(res[1]) : none());
-
-Deno.test("optionOk", () => {
-    assertEquals(optionOk(ok(2)), some(2));
-    assertEquals(optionOk(err("nothing left")), none());
-});
 
 /**
  * Takes the error value as an optional if the result is an `Err`, otherwise returns `None`.
  *
  * @param res - The source result.
  * @returns The error value if exists.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { err, ok, optionErr } from "./result.ts";
+ * import { some, none } from "./option.ts";
+ * import { assertEquals } from "../deps.ts";
+ *
+ * assertEquals(optionErr(ok(2)), none());
+ * assertEquals(optionErr(err("nothing left")), some("nothing left"));
+ * ```
  */
 export const optionErr = <E, T>(res: Result<E, T>): Option<E> =>
     isErr(res) ? some(res[1]) : none();
-
-Deno.test("optionErr", () => {
-    assertEquals(optionErr(ok(2)), none());
-    assertEquals(optionErr(err("nothing left")), some("nothing left"));
-});
 
 /**
  * Converts into a string for debug. It is not safe for serialization.
  *
  * @param opt - The result value.
  * @returns The string form.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { err, ok, toString } from "./result.ts";
+ * import { assertEquals } from "../deps.ts";
+ *
+ * assertEquals(toString(ok(24)), "ok(24)");
+ * assertEquals(toString(err("hoge")), "err(hoge)");
+ * ```
  */
 export const toString = <E, T>(res: Result<E, T>) =>
     isOk(res) ? `ok(${res[1]})` : `err(${res[1]})`;
-
-Deno.test("toString", () => {
-    assertEquals(toString(ok(24)), "ok(24)");
-    assertEquals(toString(err("hoge")), "err(hoge)");
-});
 
 /**
  * Converts into an array.
  *
  * @param opt - The result value.
  * @returns The array which contains zero or one success element.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { err, ok, toArray } from "./result.ts";
+ * import { assertEquals } from "../deps.ts";
+ *
+ * assertEquals(toArray(ok(24)), [24]);
+ * assertEquals(toArray(err("hoge")), []);
+ * ```
  */
 export const toArray = <E, T>(res: Result<E, T>) =>
     optionToArray(optionOk(res));
-
-Deno.test("toArray", () => {
-    assertEquals(toArray(ok(24)), [24]);
-    assertEquals(toArray(err("hoge")), []);
-});
 
 /**
  * Maps the function onto `Result<E, _>`.
@@ -463,6 +527,18 @@ export const biMap =
  *
  * @param optRes - `Option` containing `Result`.
  * @returns `Result` containing `Option`.
+ *
+ * # Examples
+ *
+ * ```ts
+ * import { err, ok, resOptToOptRes } from "./result.ts";
+ * import { some, none } from "./option.ts";
+ * import { assertEquals } from "../deps.ts";
+ *
+ * assertEquals(resOptToOptRes(ok(some(5))), some(ok(5)));
+ * assertEquals(resOptToOptRes(ok(none())), none());
+ * assertEquals(resOptToOptRes(err("hoge")), some(err("hoge")));
+ * ```
  */
 export const resOptToOptRes = <E, T>(
     resOpt: Result<E, Option<T>>,
@@ -475,12 +551,6 @@ export const resOptToOptRes = <E, T>(
     }
     return none();
 };
-
-Deno.test("resOptToOptRes", () => {
-    assertEquals(resOptToOptRes(ok(some(5))), some(ok(5)));
-    assertEquals(resOptToOptRes(ok(none())), none());
-    assertEquals(resOptToOptRes(err("hoge")), some(err("hoge")));
-});
 
 /**
  * Applies the function to another value on `Result`.
